Add a script for computing tx fee
